+++ src/drawings/impaledDrawing.ts
@@ -5,7 +5,7 @@ export function drawImpaled(error: CustomError | Error): void {
   const errorCode =
     "code" in error || "status" in error ? (error as CustomError).code || (error as CustomError).status : "xxx";
   const errorString = error.message || String(error);
-  console.log(error);
+
   let terminalColour = terminalColours();
   console.log(`---------------------------------------`);
   console.log(`------------${terminalColour.BgRed}ERROR CODE ${errorCode}${terminalColour.Reset}-------------`);